Add basic functionality to Leader
Up to this point, many things were being slowed down due to the fact
that the leader was empty. Now that it has most of what it will need
going forward, the Swarmer class can be largely finished now.

The main features are the swarm and formation vectors with getters, join
and leave swarm functions, and the isRecruiting flag.