Add flocking state and fix swarmer transition
The flocking state is just a basic flocking behaviour, with the addition
of using the agents leader in addition to the boids. Asside from the
change in types, everything is basicaly the same as in AIProject.

Some changes were made to Swarmer and Leader classes. The leader
functions returning its swarm and formation no longer return a
reference, but a copy. This may be costly on memory and speed, but it
makes the code for the flocking functions far easier to read. Swarmer
also gained getLeader, returning a pointer to its leader. This is
nessesary for the flocking behaviour to work without adding yet another
pointer map, and should prove useful for other things should this
program be expanded.

The swarmers transition functions have been fixed slightly. Though
testing isnt possible until the leaders state machine is setup, the
flocking transition should work.