Fix swarmer flocking, add KeyBoardBehaviour
FlockingStaeBehav is now working as intended. The issue was with the
removing of other agents from the swarm, in the case that the agent
before this would be removed, and this would be skipped and remain in
the swarm, causing the distance to be NaN in seperation. On top of this,
multiple other issues have been fixed through some testing. Along with
the flocking force, there is also a wander force to give a more random
movement to swarmers, as well as a seek toward the leader if they get
too far away.

The easiest way to test that swarmers are working with leaders is to
take control of a leader, so the keyboard controller has been added for
testing purposes.