Fix node link range, add graph drawing
The linking range when creating the graph was too short and none of the
nodes were being connected, so its been increased.

Graph2D already has a function that would draw the compleate graph, but
being private in GameManager, GameApp cant use it, so GameManager now
has a public function to allow it. It is called when a flag is true,
being toggled when tab is pressed.