Add AttackState, fix formation setup
Previously, when a swarmer was set to enter formation, it wasnt added to
the formation vector, and therefore it was concidered that there was no
formation. To fix this, the Leader class has a new addToFormation function,
adding the passed swarmer to the vector, which is ued by FindPlayerState
when triggering the formation sub-state.

The new AttackState will change the formation swarmers to follow the
player, then have the leader keep their distance from the player. A
transition has also been added, going from FindPlayerState to AttackState
when within range of the player. Initial testing shows this to work as
intended.