Add RecruitmentState and its transition
The recruitment state is very simple, set the recruiting flag to true and
dont move, then when exiting the state, set the flag to false. It does
exactly this. The transition just checks the agent's swarm size, and if
its larger than 50, it triggers.

Because this state is triggered by FindSpawnerState, which hasnt been
implemented yet, it has not been tested. However, being as simple as it
is, I expect it to work as intended.