Clean GameApp.cpp and fix FindPlayerState
GameApp has been largly thrown together as new things have been added,
but now that all agent related stuff has been added, it can, and has,
be cleaned up and properly commented.

FindPlayerState would keep an agents path even after they wernt using the
satte any more. Now it is cleared it the shutdown function. The global
pathInfo is not in a namespace after finding it to conflict with
FindSpawnerPath, as to prevent any other conflicts.