Add Player, and Spawner health variable
The player agent consists of two parts: the player base and the attack
aoe. The base is stationary and is the actual 'player' that the leader goes after, while
the attack aoe is at the mouse curser and damages swarmers inside its
range when left click is held.

The spawner now has a variable set by the constructor that swarmers's
health is set to when created. This allows the value to be set in
GameApp's startup.