Clean code, draw swarmers as rings
The code in FindPlayerState and FindSpawnerState had some poor code that
must be kept, and has been properly commented. The arrive radius for
these states has also been reduced to improve the look of path finding.

Swarmers are now drawn as rings that become thinner as their health
drops. When as full health, however, they are a full circle.